requirements.txt
rx==3.2.0
mysql-connector-python==8.0.33
psycopg2-binary==2.9.6
